+++ Sync/app/src/main/java/com/example/sync/Administrator.java
@@ -5,10 +5,8 @@ import android.view.View;
 import android.widget.AdapterView;
 import android.widget.Button;
 import android.widget.EditText;
-import android.widget.TextView;
-import androidx.appcompat.app.AppCompatActivity;
 
-public class Administrator extends User{
+public class Administrator extends User {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -29,16 +27,6 @@ public class Administrator extends User{
         Profile profile = Profile.find(profileNumber);
 
     }
-    private void editProfile(Profile editProfile){
-        profileList.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-            @Override
-            public void onItemClick(AdapterView<?> adapterView, View view, int i, long l) {
-                new EditProfileFragment((Profile)profileList.getItemAtPosition(i)).show(getSupportFragmentManager(),"Add/Edit_City");
-
-            }
-        });
-    }
-
     @Override
     public void onConfirmPressed(Profile newProfile) {
         profileAdapter.add(newProfile);
