+++ Sync/app/src/main/java/com/example/sync/Attendee.java
@@ -6,35 +6,37 @@ import android.app.NotificationManager;
 import android.app.PendingIntent;
 import android.content.Intent;
 import android.os.Build;
+import android.os.Bundle;
 import android.view.View;
-import android.widget.AdapterView;
 import android.widget.Button;
-
 
 import java.util.ArrayList;
 
 public class Attendee extends User {
     private ArrayList<Event> events;
-    public void scanQRCode(){
-        new QRCodeScanActivity();
-    }
-    private void editProfile(Profile editProfile){
-        Button editProfileButton = findViewById(R.id.edit_File_Button);
-        profileList.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-            @Override
-            public void onItemClick(AdapterView<?> adapterView, View view, int i, long l) {
-                new EditProfileFragment((Profile)profileList.getItemAtPosition(i)).show(getSupportFragmentManager(),"Add/Edit_City");
 
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_attendee); // Replace your_layout_file with the actual file name
+
+        Button setNotificationButton = findViewById(R.id.set_Notification_Button);
+        setNotificationButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                setNotification();
+            }
+        });
+
+        Button browseEventButton = findViewById(R.id.browse_event_button);
+        browseEventButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                browseEvent();
             }
         });
     }
-    Button setNotificationButton = findViewById(R.id.set_Notification_Button);
-    setNotificationButton.setOnClickListener(new View.OnClickListener() {
-        @Override
-        public void onClick(View v) {
-            setNotification();
-        }
-    });
+
     public void setNotification(){
         // Create the NotificationChannel
         NotificationManager notificationManager = getSystemService(NotificationManager.class);
@@ -55,13 +57,6 @@ public class Attendee extends User {
         long notificationTime = eventTimeInMillis - 3600000; // 1 hour before the event
         alarmManager.set(AlarmManager.RTC_WAKEUP, notificationTime, pendingIntent);
     }
-    Button browseEventButton = findViewById(R.id.browse_event_button);
-    browseEventButton.set.setOnClickListener(new View.OnClickListener() {
-        @Override
-        public void onClick(View v) {
-            browseEvent();
-        }
-    });
     public void browseEvent(){
         events = getEventInfo();
         for (Event event : events) {
